feat: Implement middleware to fetch user ID from email

+++ Swar.API/Controllers/LikedSongsController.cs
@@ -1,7 +1,6 @@
 ﻿using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Swar.API.Exceptions;
-using Swar.API.Helpers;
 using Swar.API.Interfaces.Services;
 using Swar.API.Models;
 
@@ -32,7 +31,8 @@ namespace Swar.API.Controllers
 
             try
             {
-                int userId = UserHelper.GetUserId(User);
+                if (!HttpContext.Items.TryGetValue("UserId", out var userIdObj) || userIdObj is not int userId)
+                    throw new EntityNotFoundException("User does not exist");
                 var result = await _likedSongsService.AddSongToLikedSongs(userId, songId);
 
                 return Ok(result);
@@ -66,7 +66,8 @@ namespace Swar.API.Controllers
         {
             try
             {
-                int userId = UserHelper.GetUserId(User);
+                if (!HttpContext.Items.TryGetValue("UserId", out var userIdObj) || userIdObj is not int userId)
+                    throw new EntityNotFoundException("User does not exist");
                 var result = await _likedSongsService.RemoveSongFromLikedSongs(userId, songId);
 
                 return Ok(result);
@@ -91,7 +92,8 @@ namespace Swar.API.Controllers
         {
             try
             {
-                int userId = UserHelper.GetUserId(User);
+                if (!HttpContext.Items.TryGetValue("UserId", out var userIdObj) || userIdObj is not int userId)
+                    throw new EntityNotFoundException("User does not exist");
                 var result = await _likedSongsService.GetAllLikedSongs(userId);
 
                 return Ok(result);
@@ -117,7 +119,8 @@ namespace Swar.API.Controllers
         {
             try
             {
-                int userId = UserHelper.GetUserId(User);
+                if (!HttpContext.Items.TryGetValue("UserId", out var userIdObj) || userIdObj is not int userId)
+                    throw new EntityNotFoundException("User does not exist");
                 var result = await _likedSongsService.IsSongLikedByUser(userId, songId);
 
                 return Ok(result);
