+++ Swar.API/Controllers/PlaylistController.cs
@@ -1,7 +1,6 @@
 ﻿using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Swar.API.Exceptions;
-using Swar.API.Helpers;
 using Swar.API.Interfaces.Services;
 using Swar.API.Models;
 using Swar.API.Models.DTOs;
@@ -33,7 +32,8 @@ namespace Swar.API.Controllers
 
             try
             {
-                int userId = UserHelper.GetUserId(User);
+                if (!HttpContext.Items.TryGetValue("UserId", out var userIdObj) || userIdObj is not int userId)
+                    throw new EntityNotFoundException("User does not exist");
                 var result = await _playlistService.AddPlaylist(userId, addPlaylistDTO);
 
                 return Ok(result);
@@ -59,7 +59,8 @@ namespace Swar.API.Controllers
         {
             try
             {
-                int userId = UserHelper.GetUserId(User);
+                if (!HttpContext.Items.TryGetValue("UserId", out var userIdObj) || userIdObj is not int userId)
+                    throw new EntityNotFoundException("User does not exist");
                 var result = await _playlistService.DeletePlaylist(userId, playlistId);
 
                 return Ok(result);
@@ -111,7 +112,8 @@ namespace Swar.API.Controllers
         {
             try
             {
-                int userId = UserHelper.GetUserId(User);
+                if (!HttpContext.Items.TryGetValue("UserId", out var userIdObj) || userIdObj is not int userId)
+                    throw new EntityNotFoundException("User does not exist");
                 var result = await _playlistService.GetAllPlaylistsByUserId(userId);
 
                 return Ok(result);
@@ -137,7 +139,8 @@ namespace Swar.API.Controllers
         {
             try
             {
-                int userId = UserHelper.GetUserId(User);
+                if (!HttpContext.Items.TryGetValue("UserId", out var userIdObj) || userIdObj is not int userId)
+                    throw new EntityNotFoundException("User does not exist");
                 var result = await _playlistService.GetPlaylistById(userId, playlistId);
 
                 return Ok(result);
@@ -168,7 +171,8 @@ namespace Swar.API.Controllers
         {
             try
             {
-                int userId = UserHelper.GetUserId(User);
+                if (!HttpContext.Items.TryGetValue("UserId", out var userIdObj) || userIdObj is not int userId)
+                    throw new EntityNotFoundException("User does not exist");
                 var result = await _playlistService.UpdatePlaylist(userId, playlistId, updatePlaylistDTO);
 
                 return Ok(result);
@@ -199,7 +203,8 @@ namespace Swar.API.Controllers
         {
             try
             {
-                int userId = UserHelper.GetUserId(User);
+                if (!HttpContext.Items.TryGetValue("UserId", out var userIdObj) || userIdObj is not int userId)
+                    throw new EntityNotFoundException("User does not exist");
                 var result = await _playlistService.UpdatePlaylistPrivacy(userId, playlistId, updatePlaylistPrivacyDTO.IsPrivate);
 
                 return Ok(result);
