+++ Swar.API/Controllers/PlaylistSongsController.cs
@@ -1,7 +1,6 @@
 ﻿using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Swar.API.Exceptions;
-using Swar.API.Helpers;
 using Swar.API.Interfaces.Services;
 using Swar.API.Models;
 using Swar.API.Models.DTOs;
@@ -33,7 +32,8 @@ namespace Swar.API.Controllers
 
             try
             {
-                int userId = UserHelper.GetUserId(User);
+                if (!HttpContext.Items.TryGetValue("UserId", out var userIdObj) || userIdObj is not int userId)
+                    throw new EntityNotFoundException("User does not exist");
                 var result = await _playlistSongsService.AddSongToPlaylist(userId, playlistSongsDTO);
 
                 return Ok(result);
@@ -68,7 +68,8 @@ namespace Swar.API.Controllers
         {
             try
             {
-                int userId = UserHelper.GetUserId(User);
+                if (!HttpContext.Items.TryGetValue("UserId", out var userIdObj) || userIdObj is not int userId)
+                    throw new EntityNotFoundException("User does not exist");
                 var result = await _playlistSongsService.RemoveSongFromPlaylist(userId, playlistId, songId);
 
                 return Ok(result);
@@ -120,7 +121,8 @@ namespace Swar.API.Controllers
         {
             try
             {
-                int userId = UserHelper.GetUserId(User);
+                if (!HttpContext.Items.TryGetValue("UserId", out var userIdObj) || userIdObj is not int userId)
+                    throw new EntityNotFoundException("User does not exist");
                 var result = await _playlistSongsService.GetAllSongsInUserPlaylist(userId, publicId);
 
                 return Ok(result);
